Delete README.md
Application to bridge the communication gap between people with speaking, hearing disabilities and with normal people. Used the technology of hand gesture recognition. By adding certain images of gestures we can train the software through tensor flow such that whenever it matches gesture to the stored one it will perform the set of instructions under it.